+++ src/components/Clock.tsx
@@ -1,0 +1,16 @@
+import { useEffect, useState } from 'react'
+
+export const Clock = () => {
+  const [time, setTime] = useState(new Date())
+
+  useEffect(() => {
+    const timer = setInterval(() => setTime(new Date()), 1000)
+    return () => clearInterval(timer)
+  }, [])
+
+  const formatTime = (date: Date) => {
+    return `${date.getHours().toString().padStart(2, '0')}:${date.getMinutes().toString().padStart(2, '0')}`
+  }
+
+  return <span>{formatTime(time)}</span>
+}
