+++ src/components/ImageSlideComponent.tsx
@@ -1,0 +1,9 @@
+import type { ImageSlideData } from '../types'
+
+export const ImageSlideComponent = ({
+  content,
+}: { content: ImageSlideData }) => (
+  <div className="relative z-40 h-full w-full bg-black">
+    <img src={content.url} alt="" className="h-full w-full object-cover" />
+  </div>
+)
